decouple gist loading and layout logic, add editor page
