+++ package.json
@@ -24,6 +24,7 @@
     "backbone.modal": "^1.1.5",
     "bluebird": "^3.1.1",
     "bootstrap": "^3.3.5",
+    "brace": "^0.7.0",
     "chroma-js": "^1.1.1",
     "clipboard": "^1.4.2",
     "datatables": "https://github.com/DataTables/DataTables/archive/1.10.9.tar.gz",
@@ -35,6 +36,7 @@
     "numbro": "^1.3.3",
     "soda-js": "socrata/soda-js",
     "spin.js": "^2.3.2",
+    "split.js": "^1.0.6",
     "underscore": "^1.8.3"
   },
   "devDependencies": {
