+++ src/editor.html
@@ -1,0 +1,93 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>VizWit Editor</title>
+
+		
+		<link rel="stylesheet" href="//fonts.googleapis.com/css?family=Arimo:400,700,400italic">
+		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css">
+		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
+		<link rel="stylesheet" href="https://cdn.datatables.net/1.10.9/css/dataTables.bootstrap.min.css">
+		<link rel="stylesheet" href="https://api.mapbox.com/mapbox.js/v2.2.1/mapbox.css">
+		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/gridstack.js/0.2.3/gridstack.min.css">
+		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/backbone.modal/1.1.5/backbone.modal-min.css">
+		<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/backbone.modal/1.1.5/backbone.modal.theme-min.css">
+		<link rel="stylesheet" href="styles/main.css">
+
+    <style type="text/css">
+    html, body, .row { height: 100%; }
+    body {
+      box-sizing: border-box;
+      margin: 0;
+    }
+    
+    .preview-container { padding: 10px; }
+
+    .gutter {
+      background-color: #eee;
+      background-repeat: no-repeat;
+      background-position: 50%;
+    }
+
+    .gutter.gutter-horizontal {
+      background-image: url('https://raw.githubusercontent.com/nathancahill/Split.js/master/grips/vertical.png');
+      cursor: ew-resize;
+    }
+
+    .gutter.gutter-vertical {
+      background-image: url('https://raw.githubusercontent.com/nathancahill/Split.js/master/grips/horizontal.png');
+      cursor: ns-resize;
+    }
+
+    .split {
+      -webkit-box-sizing: border-box;
+        -moz-box-sizing: border-box;
+              box-sizing: border-box;
+    }
+
+    .split, .gutter.gutter-horizontal {
+      height: 100%;
+      float: left;
+    }
+    </style>
+  </head>
+  <body>
+    <div class="row">
+      <div id="editor" class="split">{
+	"version": "2",
+	"header": {
+		"title": "Sample Page",
+		"description": "This is a sample page for the editor tool"
+	},
+	"cards": [{
+		"x": 0,
+		"y": 0,
+		"width": 6,
+		"height": 4,
+		"vizwit": {
+			"title": "Date",
+			"chartType": "datetime",
+			"chartType2": "line",
+			"domain": "data.cityofchicago.org",
+			"dataset": "6zsd-86xi",
+			"groupBy": "date_trunc_ym(date)",
+			"triggerField": "date"
+		}
+	}]
+}
+      </div>
+      <div id="preview" class="split">
+        <div class="header-container">
+          <div class="preview-container" id="page-header"></div>
+        </div>
+        
+        <div class="preview-container">
+          <div class="grid-stack grid-stack-12" id="page-content"></div>
+        </div>
+      </div>
+    </div>
+
+    <script src="scripts/editor.js"></script>
+  </body>
+</html>
