+++ src/index.html
@@ -42,6 +42,6 @@
 			<div class="grid-stack grid-stack-12" id="page-content"></div>
 		</div>
 		
-		<script src="scripts/layout.js"></script>
+		<script src="scripts/layout-from-gist.js"></script>
 	</body>
 </html>
